OSIV 설정 false로 하기 위한 리팩토링

+++ src/main/java/com/webtoon/cartoon/domain/Cartoon.java
@@ -105,7 +105,7 @@ public class Cartoon extends BaseTimeEntity {
         this.likes += 1;
     }
 
-    public double calculateSum(int cartoonListSize) {
+    public double calculateSum(long cartoonListSize) {
         return (cartoonListSize * rating);
     }
